Push the definition of the `Expr` datatype into the `What4.Expr.App` module, and change the traversal operation on `App` so that it doesn't descend into the body of defined functions.
